# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "API documentation for Laravel application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tenants": {
            "get": {
                "summary": "List all tenants",
                "operationId": "9f953e00ff6b37d7bd59604347891b9a",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Create a new tenant",
                "operationId": "d7c436dd05ddff8abafab0bf3d7c76ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "domain",
                                    "plan_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "domain": {
                                        "type": "string"
                                    },
                                    "plan_id": {
                                        "type": "integer"
                                    },
                                    "active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tenant created successfully"
                    }
                }
            }
        },
        "/api/tenants/{id}": {
            "get": {
                "summary": "Get a specific tenant",
                "operationId": "6334a1aea0c9dc312536a5744e06d805",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Tenant not found"
                    }
                }
            },
            "put": {
                "summary": "Update a tenant",
                "operationId": "f70783d49d2b97145271b2237621e1ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "domain",
                                    "plan_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "domain": {
                                        "type": "string"
                                    },
                                    "plan_id": {
                                        "type": "integer"
                                    },
                                    "active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tenant updated successfully"
                    },
                    "404": {
                        "description": "Tenant not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a tenant",
                "operationId": "69e237daa88f624a196e9ce27b975e21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant deleted successfully"
                    },
                    "404": {
                        "description": "Tenant not found"
                    }
                }
            }
        }
    }
}